requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
Django==4.2.1
django-crispy-forms==2.0
django-modeltranslation==0.18.9
idna==3.4
psycopg2-binary==2.9.6
python-decouple==3.8
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.6.2
urllib3==2.0.2
